+++ custom_components/cover_time_based/websocket_api.py
@@ -67,6 +67,8 @@ from .cover import (
     DEFAULT_REPORTS_COMMAND_NOT_ENDPOINT,
     DEFAULT_SEND_ENDPOINT_STOP,
     DEFAULT_WAIT_FOR_RELAY_FEEDBACK,
+    MIN_DURATION,
+    PERCENT,
 )
 from .helpers import resolve_entity_or_none
 
@@ -284,7 +286,7 @@ async def ws_get_config(
             ]
         ),
         vol.Optional("pulse_time"): vol.Any(
-            None, vol.All(vol.Coerce(float), vol.Range(min=0.1, max=10))
+            None, vol.All(vol.Coerce(float), vol.Range(min=MIN_DURATION, max=10))
         ),
         vol.Optional("relay_reports_off"): vol.Any(None, bool),
         vol.Optional("send_endpoint_stop"): vol.Any(None, bool),
@@ -313,16 +315,16 @@ async def ws_get_config(
             ]
         ),
         vol.Optional("travel_time_close"): vol.Any(
-            None, vol.All(vol.Coerce(float), vol.Range(min=0.1, max=600))
+            None, vol.All(vol.Coerce(float), vol.Range(min=MIN_DURATION, max=600))
         ),
         vol.Optional("travel_time_open"): vol.Any(
-            None, vol.All(vol.Coerce(float), vol.Range(min=0.1, max=600))
+            None, vol.All(vol.Coerce(float), vol.Range(min=MIN_DURATION, max=600))
         ),
         vol.Optional("tilt_time_close"): vol.Any(
-            None, vol.All(vol.Coerce(float), vol.Range(min=0.1, max=600))
+            None, vol.All(vol.Coerce(float), vol.Range(min=MIN_DURATION, max=600))
         ),
         vol.Optional("tilt_time_open"): vol.Any(
-            None, vol.All(vol.Coerce(float), vol.Range(min=0.1, max=600))
+            None, vol.All(vol.Coerce(float), vol.Range(min=MIN_DURATION, max=600))
         ),
         vol.Optional("travel_startup_delay"): vol.Any(
             None, vol.All(vol.Coerce(float), vol.Range(min=0, max=600))
@@ -341,12 +343,8 @@ async def ws_get_config(
         vol.Optional("direction_change_delay"): vol.Any(
             None, vol.All(vol.Coerce(float), vol.Range(min=0, max=600))
         ),
-        vol.Optional("safe_tilt_position"): vol.Any(
-            None, vol.All(vol.Coerce(int), vol.Range(min=0, max=100))
-        ),
-        vol.Optional("max_tilt_allowed_position"): vol.Any(
-            None, vol.All(vol.Coerce(int), vol.Range(min=0, max=100))
-        ),
+        vol.Optional("safe_tilt_position"): vol.Any(None, PERCENT),
+        vol.Optional("max_tilt_allowed_position"): vol.Any(None, PERCENT),
         vol.Optional("tilt_open_switch"): vol.Any(str, None),
         vol.Optional("tilt_close_switch"): vol.Any(str, None),
         vol.Optional("tilt_stop_switch"): vol.Any(str, None),